requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.10
pydantic==2.9.2
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3
textblob==0.18.0
python-multipart==0.0.12
nltk==3.9.1
scipy==1.14.1